requirements.txt
fastapi
pydantic
uvicorn[standard]
python-multipart
numpy
Pillow
tensorflow
google-cloud-storage
httpx
requests
opencv-python
python-dotenv
google-auth
h5py
scikit-learn